moved add_validator test
